+++ azext_iot/adr/commands_link.py
@@ -134,32 +134,32 @@ def adr_link_dps_list(cmd, namespace_name: str, resource_group_name: str):
     )
 
 
-# ==================== link adu ====================
+# ==================== link du ====================
 
 
-def adr_link_adu_add(
+def adr_link_du_add(
     cmd,
     endpoint_name: str,
     namespace_name: str,
     resource_group_name: str,
-    adu_resource_id: str,
+    du_resource_id: str,
     mi_system_assigned: bool = False,
     mi_user_assigned: Optional[str] = None,
     no_wait: bool = False,
 ):
     provider = LinkProvider(cmd)
-    return provider.adu_add(
+    return provider.du_add(
         endpoint_name=endpoint_name,
         namespace_name=namespace_name,
         resource_group_name=resource_group_name,
-        adu_resource_id=adu_resource_id,
+        du_resource_id=du_resource_id,
         mi_system_assigned=mi_system_assigned,
         mi_user_assigned=mi_user_assigned,
         no_wait=no_wait,
     )
 
 
-def adr_link_adu_update(
+def adr_link_du_update(
     cmd,
     endpoint_name: str,
     namespace_name: str,
@@ -169,7 +169,7 @@ def adr_link_adu_update(
     no_wait: bool = False,
 ):
     provider = LinkProvider(cmd)
-    return provider.adu_update(
+    return provider.du_update(
         endpoint_name=endpoint_name,
         namespace_name=namespace_name,
         resource_group_name=resource_group_name,
@@ -179,18 +179,18 @@ def adr_link_adu_update(
     )
 
 
-def adr_link_adu_show(cmd, endpoint_name: str, namespace_name: str, resource_group_name: str):
+def adr_link_du_show(cmd, endpoint_name: str, namespace_name: str, resource_group_name: str):
     provider = LinkProvider(cmd)
-    return provider.adu_show(
+    return provider.du_show(
         endpoint_name=endpoint_name,
         namespace_name=namespace_name,
         resource_group_name=resource_group_name,
     )
 
 
-def adr_link_adu_list(cmd, namespace_name: str, resource_group_name: str):
+def adr_link_du_list(cmd, namespace_name: str, resource_group_name: str):
     provider = LinkProvider(cmd)
-    return provider.adu_list(
+    return provider.du_list(
         namespace_name=namespace_name,
         resource_group_name=resource_group_name,
     )